docker_project/requirements.txt
Flask
python-dotenv
langchain
langchain_core
langchain_community
langgraph
openai
numpy
pandas
matplotlib
langchain_openai
requests
